RavenDB-13107 - PR comments
