+++ src/Raven.Server/Documents/Handlers/QueriesHandler.cs
@@ -506,16 +506,6 @@ namespace Raven.Server.Documents.Handlers
         /// TrafficWatchQuery writes query data to httpContext
         /// </summary>
         /// <param name="indexQuery"></param>
-        private void TrafficWatchQuery(IndexQueryServerSide indexQuery)
-        {
-            var sb = new StringBuilder();
-            // append stringBuilder with the query
-            sb.Append(indexQuery.Query);
-            // if query got parameters append with parameters
-            if (indexQuery.QueryParameters != null && indexQuery.QueryParameters.Count > 0)
-                sb.AppendLine().Append(indexQuery.QueryParameters);
-            AddStringToHttpContext(sb.ToString(), TrafficWatchChangeType.Queries);
-        }
 
         private async Task ExecuteQueryOperation(IndexQueryServerSide query,
                 Func<QueryRunner,
