+++ src/Raven.Server/Documents/Sharding/ShardedContext.cs
@@ -1,47 +1,39 @@
 ﻿using System;
-using System.Collections.Concurrent;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using Raven.Client;
 using Raven.Client.Documents.Conventions;
-using Raven.Client.Documents.Indexes;
 using Raven.Client.Http;
 using Raven.Client.ServerWide;
-using Raven.Server.Config;
-using Raven.Server.Documents.Indexes;
-using Raven.Server.Documents.Indexes.Static;
-using Raven.Server.Documents.Patch;
 using Raven.Server.Documents.Queries;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Utils;
 using Sparrow;
+using Sparrow.Json;
 using Sparrow.Server;
 using Sparrow.Utils;
+using Voron;
 
 namespace Raven.Server.Documents.Sharding
 {
-    public unsafe partial class ShardedContext
+    public partial class ShardedContext
     {
         public const int NumberOfShards = 1024 * 1024;
 
-        private readonly ServerStore _server;
-        public QueryMetadataCache QueryMetadataCache = new QueryMetadataCache();
-
         private readonly ServerStore _serverStore;
+        private readonly long _lastClientConfigurationIndex;
+        private readonly ShardedIndexesCache _indexes;
+        private readonly ShardExecutor _shardExecutor;
+
         private DatabaseRecord _record;
         public RequestExecutor[] RequestExecutors;
-        private readonly long _lastClientConfigurationIndex;
+        public QueryMetadataCache QueryMetadataCache = new();
 
-        private readonly ConcurrentDictionary<string, IndexDefinition> _cachedMapReduceIndexDefinitions = new(StringComparer.OrdinalIgnoreCase);
-        private readonly ConcurrentDictionary<string, AutoIndexDefinition> _cachedAutoMapReduceIndexDefinitions = new(StringComparer.OrdinalIgnoreCase);
-        public readonly ScriptRunnerCache ScriptRunnerCache;
-        private readonly RavenConfiguration _configuration;
-
-        private ShardExecutor _shardExecutor;
         public ShardExecutor ShardExecutor => _shardExecutor;
         public DatabaseRecord DatabaseRecord => _record;
+
+        public ShardedIndexesCache Indexes => _indexes;
 
         public int[] FullRange;
 
@@ -52,23 +44,7 @@ namespace Raven.Server.Documents.Sharding
 
             _serverStore = serverStore;
             _record = record;
-
-            UpdateMapReduceIndexes(record.Indexes
-                .Where(x => x.Value.Type.IsStaticMapReduce())
-                .ToDictionary(x => x.Key, x => x.Value));
-
-            UpdateAutoMapReduceIndexes(record.AutoIndexes
-                .Where(x => x.Value.Type.IsAutoMapReduce())
-                .ToDictionary(x => x.Key, x => x.Value));
-
-            _configuration = RavenConfiguration.CreateForDatabase(_serverStore.Configuration, DatabaseName);
-
-            foreach ((string key, string value) in _record.Settings)
-                _configuration.SetSetting(key, value);
-
-            _configuration.Initialize();
-
-            ScriptRunnerCache = new ScriptRunnerCache(database: null, _configuration);
+            _indexes = new ShardedIndexesCache(serverStore, record);
 
             _lastClientConfigurationIndex = serverStore.LastClientConfigurationIndex;
 
@@ -90,7 +66,7 @@ namespace Raven.Server.Documents.Sharding
             Streaming = new ShardedStreaming(this);
         }
 
-        public IDisposable AllocateContext(out JsonOperationContext context) => _server.ContextPool.AllocateOperationContext(out context);
+        public IDisposable AllocateContext(out JsonOperationContext context) => _serverStore.ContextPool.AllocateOperationContext(out context);
 
         public void UpdateDatabaseRecord(DatabaseRecord record)
         {
@@ -118,20 +94,25 @@ namespace Raven.Server.Documents.Sharding
         {
             using (DocumentIdWorker.GetLowerIdSliceAndStorageKey(context, key, out var lowerId, out _))
             {
-                byte* buffer = lowerId.Content.Ptr;
-                int size = lowerId.Size;
-
-                AdjustAfterSeparator((byte)'$', ref buffer, ref size);
-
-                if (size == 0)
-                    throw new ArgumentException("Key '" + key + "', has a shard id length of 0");
-
-                var hash = Hashing.XXHash64.Calculate(buffer, (ulong)size);
-                return (int)(hash % NumberOfShards);
+                return GetShardId(lowerId);
             }
         }
 
-        private static void AdjustAfterSeparator(byte expected, ref byte* ptr, ref int len)
+        private static unsafe int GetShardId(Slice lowerId)
+        {
+            byte* buffer = lowerId.Content.Ptr;
+            int size = lowerId.Size;
+
+            AdjustAfterSeparator((byte)'$', ref buffer, ref size);
+
+            if (size == 0)
+                throw new ArgumentException("Key '" + lowerId + "', has a shard id length of 0");
+
+            var hash = Hashing.XXHash64.Calculate(buffer, (ulong)size);
+            return (int)(hash % NumberOfShards);
+        }
+
+        private static unsafe void AdjustAfterSeparator(byte expected, ref byte* ptr, ref int len)
         {
             for (int i = len - 1; i > 0; i--)
             {
@@ -154,16 +135,16 @@ namespace Raven.Server.Documents.Sharding
             return _record.ShardAllocations[^1].Shard;
         }
 
+        public int GetShardIndex(Slice lowerId)
+        {
+            var shardId = GetShardId(lowerId);
+            return GetShardIndex(shardId);
+        }
+
         public int GetShardIndex(TransactionOperationContext context, string key)
         {
             var shardId = GetShardId(context, key);
-            for (int i = 0; i < _record.ShardAllocations.Count - 1; i++)
-            {
-                if (shardId < _record.ShardAllocations[i + 1].RangeStart)
-                    return _record.ShardAllocations[i].Shard;
-            }
-
-            return _record.ShardAllocations[^1].Shard;
+            return GetShardIndex(shardId);
         }
 
         public int GetShardIndex(ByteStringContext context, LazyStringValue key)
@@ -189,62 +170,6 @@ namespace Raven.Server.Documents.Sharding
             var lastClientConfigurationIndex = _record.Client?.Etag ?? 0;
             var actual = Hashing.Combine(lastClientConfigurationIndex, _lastClientConfigurationIndex);
             return actual > clientConfigurationEtag;
-        }
-
-        public void UpdateIndexes(RawDatabaseRecord record)
-        {
-            UpdateMapReduceIndexes(record.MapReduceIndexes());
-            UpdateAutoMapReduceIndexes(record.AutoMapReduceIndexes());
-        }
-
-        private void UpdateMapReduceIndexes(Dictionary<string, IndexDefinition> indexDefinitions)
-        {
-            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Grisha, DevelopmentHelper.Severity.Normal,
-                "remove the deleted map reduce indexes from cache");
-
-            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Grisha, DevelopmentHelper.Severity.Major,
-                "handle side-by-side");
-
-            foreach ((string indexName, IndexDefinition definition) in indexDefinitions)
-            {
-                Debug.Assert(definition.Type.IsStaticMapReduce());
-
-                _cachedMapReduceIndexDefinitions[indexName] = definition;
-            }
-        }
-
-        private void UpdateAutoMapReduceIndexes(Dictionary<string, AutoIndexDefinition> autoIndexDefinitions)
-        {
-            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Grisha, DevelopmentHelper.Severity.Normal,
-                "remove the deleted auto map-reduce indexes from cache");
-
-            foreach ((string indexName, AutoIndexDefinition definition) in autoIndexDefinitions)
-            {
-                Debug.Assert(definition.Type.IsAutoMapReduce());
-
-                _cachedAutoMapReduceIndexDefinitions[indexName] = definition;
-            }
-        }
-
-        public AbstractStaticIndexBase GetCompiledMapReduceIndex(string indexName, TransactionOperationContext context)
-        {
-            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Grisha, DevelopmentHelper.Severity.Normal, 
-                "cache the compiled JavaScript indexes - concurrent queue since they are single threaded");
-
-            return _cachedMapReduceIndexDefinitions.TryGetValue(indexName, out var indexDefinition) == false 
-                ? null 
-                : IndexCompilationCache.GetIndexInstance(indexDefinition, _configuration, IndexDefinitionBaseServerSide.IndexVersion.CurrentVersion);
-        }
-
-        public bool TryGetAutoIndexDefinition(string indexName, out AutoIndexDefinition autoIndexDefinition)
-        {
-            return _cachedAutoMapReduceIndexDefinitions.TryGetValue(indexName, out autoIndexDefinition);
-        }
-
-        public bool IsMapReduceIndex(string indexName)
-        {
-            return _cachedMapReduceIndexDefinitions.TryGetValue(indexName, out _)
-                || _cachedAutoMapReduceIndexDefinitions.TryGetValue(indexName, out _);
         }
     }
 }
